Begin switching to interactive-app
