+++ ts/examples/spelunker/package.json
@@ -30,7 +30,7 @@
     "aiclient": "workspace:*",
     "code-processor": "workspace:*",
     "dotenv": "^16.3.1",
-    "readline-sync": "^1.4.10",
+    "interactive-app": "workspace:*",
     "typeagent": "workspace:*",
     "typechat": "^0.1.1",
     "typescript": "^5.4.2"
